+++ test/unit/pouch-db-integration.test.ts
@@ -42,7 +42,6 @@ config.parallel('pouch-db-integration.test.js', () => {
             addPouchPlugin(require('pouchdb-adapter-memory'));
             addPouchPlugin(PouchDBFind);
 
-            console.log('--- 1');
             const pouch1: PouchDBInstance = new PouchDB('foobar1' + randomCouchString(10), {
                 adapter: 'memory'
             });
@@ -65,7 +64,6 @@ config.parallel('pouch-db-integration.test.js', () => {
             }
 
             // insert
-            console.log('--- 2');
             const insertResult = await pouch1.put(clone(docData));
             docData._rev = insertResult.rev;
 
@@ -83,7 +81,6 @@ config.parallel('pouch-db-integration.test.js', () => {
                 const allDocs = await getDocsFromPouch2();
                 return allDocs.length === 1 && (allDocs[0] as any).value === 2;
             });
-            console.log('--- 3');
 
             // delete
             docData._deleted = true;
@@ -95,26 +92,20 @@ config.parallel('pouch-db-integration.test.js', () => {
             });
 
             // undelete via insert
-            console.log('--- 4');
             const undeleteResult = await pouch1.put({
                 _id: 'syncMe',
                 value: 5
             });
             assert.ok(undeleteResult);
-            console.log('--- 5');
             await AsyncTestUtil.waitUntil(async () => {
                 const allDocs = await getDocsFromPouch2();
                 return allDocs.length === 1 && (allDocs[0] as any).value === 5;
             });
 
-            console.log('--- 6');
             await syncHandler.cancel();
             await promiseWait(100);
-            console.log('--- 6.2');
             await pouch1.close();
-            console.log('--- 6.5');
             await pouch2.close();
-            console.log('--- 7');
         });
         it('must be able to insert->update->delete via new_edits:false', async () => {
             const pouch: PouchDBInstance = new PouchDB('foobar' + randomCouchString(10), {